c305: extract RepresentationFitTab + extract TOPIC_FAMILY_REPS to shared workspace/topicFamilyReps.ts (#441 2.1)
Workspace.tsx 6169 LOC -> 5926 LOC (-243 LOC). Entry chunk 163.41 KB -> 155.88 KB (-7.5 KB).
23 tabs extracted + lazy.
Removed unused Scatter3D lazy import from Workspace.tsx (it moves with RepresentationFitTab).
typecheck clean, 44/44 tests pass.
