+++ frontend/src/pages/workspace/topicFamilyReps.ts
@@ -1,0 +1,9 @@
+export const TOPIC_FAMILY_REPS = new Set([
+  "lda",
+  "lda_sparse",
+  "lda_tomo",
+  "hdp",
+  "ctm",
+  "prodlda",
+  "etm",
+]);
